requirements.txt
Flask==2.1.2
python-dotenv==0.20.0
SQLAlchemy==1.4.37
python-dateutil==2.8.2
bcrypt==3.2.2
PyJWT==2.4.0